Try to protect against setting or adding non-values to metrics
Its not hard in metrics proxy/bridge code to accidentally pass through
an undef as a metric value because something has happened at the remote
metrics source.

This is definitely a bug in the client code, which should be trying to
sanitise incoming data, but we can at least try to help make these bugs
easier to see by warning for the simple cases, and not breaking the
metrics output (leading to failed scrapes).